fix: support recovery from partial deletes (#53)
Add a tag-only deletion option for releases whose tag remains.

Report the recovery command when tag deletion fails.

+++ src/Command/DeleteRelease.php
@@ -12,6 +12,7 @@ use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
@@ -36,6 +37,11 @@ class DeleteRelease extends BaseCommand
             ->addArgument(
                 'version', InputArgument::OPTIONAL,
                 'The version of the release to delete.',
+            )
+            ->addOption(
+                'tag-only', null,
+                InputOption::VALUE_NONE,
+                'Delete only the Git tag associated with the version.',
             );
     }
 
@@ -49,6 +55,11 @@ class DeleteRelease extends BaseCommand
               The <info>version</info> argument identifies the release to delete. If it is not
               specified, you will be prompted to select from the available releases when
               running interactively. Use the <info>%s</info> command to view the available releases.
+
+            <comment>Tag only</comment>
+              Use <info>--tag-only</info> to delete only the Git tag. This is useful for recovering
+              from an incomplete release deletion. The <info>version</info> argument is required
+              when using this option.
             HELP,
             ListReleases::NAME,
         );
@@ -63,7 +74,7 @@ class DeleteRelease extends BaseCommand
     {
         parent::interact($input, $output);
 
-        if (null !== $input->getArgument('version')) {
+        if (null !== $input->getArgument('version') || $input->getOption('tag-only')) {
             return;
         }
 
@@ -117,7 +128,7 @@ class DeleteRelease extends BaseCommand
         /** @var ?string */
         $versionArg = $input->getArgument('version');
         if (null === $versionArg) {
-            throw new RuntimeException('Specify the version to delete when running non-interactively.');
+            throw new RuntimeException('Specify the version to delete when running non-interactively or using --tag-only.');
         }
 
         try {
@@ -126,20 +137,34 @@ class DeleteRelease extends BaseCommand
             throw new RuntimeException(sprintf('Invalid version "%s": %s', $versionArg, $e->getMessage()), previous: $e);
         }
 
-        $question = new ConfirmationQuestion(sprintf(
-            'You are about to delete release "%s" and its associated Git tag. Do you want to continue? (y/N)',
-            $version,
-        ), false);
+        /** @var bool */
+        $tagOnly = $input->getOption('tag-only');
+        $question = new ConfirmationQuestion(
+            $tagOnly
+                ? sprintf('You are about to delete Git tag "%s". Do you want to continue? (y/N)', $version)
+                : sprintf('You are about to delete release "%s" and its associated Git tag. Do you want to continue? (y/N)', $version),
+            false,
+        );
         if ($input->isInteractive() && !(new QuestionHelper())->ask($input, $output, $question)) {
             $output->writeln('Aborting.');
 
             return self::ABORTED;
         }
 
-        $this->gitHubClient->deleteRelease($repository, $version);
-        $output->writeln(sprintf('Successfully deleted release <info>%s</info>', $version));
+        if (!$tagOnly) {
+            $this->gitHubClient->deleteRelease($repository, $version);
+            $output->writeln(sprintf('Successfully deleted release <info>%s</info>', $version));
+        }
 
-        $this->gitHubClient->deleteTag($repository, $version);
+        try {
+            $this->gitHubClient->deleteTag($repository, $version);
+        } catch (RuntimeException $e) {
+            if (!$tagOnly) {
+                throw new RuntimeException(sprintf('Failed to delete tag "%s" after deleting its GitHub release. The release was deleted, but the tag remains. Retry with: imbo-releaser delete --tag-only %s', $version, $version), previous: $e);
+            }
+
+            throw $e;
+        }
         $output->writeln(sprintf('Successfully deleted tag <info>%s</info>', $version));
 
         return self::SUCCESS;
